move shared ROL functions to base class
